requirements.txt
fastapi
uvicorn
pydantic[dotenv]
pydantic[email]
pymongo
python-dotenv
python_bcrypt==0.3.2
python-multipart